For arrow endpoints introduce project.cypher and project.native
