+++ tests/integration/procedure_surface/arrow/catalog/test_catalog_arrow_endpoints.py
@@ -10,12 +10,9 @@ from graphdatascience.graph.graph_api import Graph
 from graphdatascience.procedure_surface.api.catalog.catalog_endpoints import (
     RelationshipPropertySpec,
 )
-from graphdatascience.procedure_surface.arrow.catalog import ProjectionResult
 from graphdatascience.procedure_surface.arrow.catalog.catalog_arrow_endpoints import (
     CatalogArrowEndpoints,
 )
-from graphdatascience.query_runner.query_runner import QueryRunner
-from graphdatascience.query_runner.query_type import QueryType
 from tests.integration.procedure_surface.arrow.graph_creation_helper import create_graph
 
 
@@ -81,99 +78,6 @@ def test_drop(catalog_endpoints: CatalogArrowEndpoints, sample_graph: Graph) -> 
 def test_drop_nonexistent(catalog_endpoints: CatalogArrowEndpoints) -> None:
     with pytest.raises(ArrowKeyError, match="does not exist on database"):
         catalog_endpoints.drop("nonexistent", fail_if_missing=True)
-
-
-@pytest.mark.db_integration
-def test_projection(arrow_client: AuthenticatedArrowClient, query_runner: QueryRunner) -> None:
-    try:
-        endpoints = CatalogArrowEndpoints(arrow_client, query_runner)
-        G, result = endpoints.project(
-            graph_name="g",
-            query="UNWIND range(1, 10) AS x WITH gds.graph.project.remote(x, null) as g RETURN g",
-        )
-
-        assert isinstance(result, ProjectionResult)
-
-        assert G.name() == "g"
-        assert result.graph_name == "g"
-        assert result.node_count == 10
-        assert result.relationship_count == 0
-        assert result.project_millis >= 0
-
-        assert len(endpoints.list("g")) == 1
-    finally:
-        endpoints.drop("g", fail_if_missing=False)
-
-
-@pytest.mark.db_integration
-def test_projection_with_query_parameters(arrow_client: AuthenticatedArrowClient, query_runner: QueryRunner) -> None:
-    try:
-        endpoints = CatalogArrowEndpoints(arrow_client, query_runner)
-        G, result = endpoints.project(
-            graph_name="g",
-            query="UNWIND range(1, $LIMIT) AS x WITH gds.graph.project.remote(x, null) as g RETURN g",
-            query_parameters={"LIMIT": 10},
-        )
-
-        assert isinstance(result, ProjectionResult)
-
-        assert G.name() == "g"
-        assert result.graph_name == "g"
-        assert result.node_count == 10
-        assert result.relationship_count == 0
-        assert result.project_millis >= 0
-
-        assert len(endpoints.list("g")) == 1
-    finally:
-        endpoints.drop("g", fail_if_missing=False)
-
-
-@pytest.mark.db_integration
-def test_projection_with_query_rewrite(arrow_client: AuthenticatedArrowClient, query_runner: QueryRunner) -> None:
-    try:
-        endpoints = CatalogArrowEndpoints(arrow_client, query_runner)
-        G, result = endpoints.project(
-            graph_name="g",
-            query="UNWIND range(1, 10) AS x WITH gds.graph.project(x, null) as g RETURN g",
-        )
-
-        assert isinstance(result, ProjectionResult)
-
-        assert G.name() == "g"
-        assert result.graph_name == "g"
-        assert result.node_count == 10
-        assert result.relationship_count == 0
-        assert result.project_millis >= 0
-
-        assert len(endpoints.list("g")) == 1
-    finally:
-        endpoints.drop("g", fail_if_missing=False)
-
-
-@pytest.mark.db_integration
-def test_store_projection(arrow_client: AuthenticatedArrowClient, query_runner: QueryRunner) -> None:
-    try:
-        query_runner.run_cypher(
-            "UNWIND range(1, 5) AS x CREATE (:Person)-[:KNOWS]->(:Person)",
-            QueryType.USER_ACTION,
-        )
-
-        endpoints = CatalogArrowEndpoints(arrow_client, query_runner)
-        G, result = endpoints.project_native(
-            graph_name="g",
-            node_label_filter=["Person"],
-            relationship_type_filter=["KNOWS"],
-        )
-
-        assert G.name() == "g"
-        assert result.graph_name == "g"
-        assert result.node_count == 10
-        assert result.relationship_count == 5
-
-        assert len(endpoints.list("g")) == 1
-    finally:
-        endpoints.drop("g", fail_if_missing=False)
-        query_runner.run_cypher("MATCH (n) DETACH DELETE n", QueryType.USER_ACTION)
 
 
 def test_construct(arrow_client: AuthenticatedArrowClient) -> None:
@@ -255,63 +159,6 @@ def test_graph_generate_with_relationships_property(catalog_endpoints: CatalogAr
         assert result.relationship_property == RelationshipPropertySpec.fixed("weight", 42)
 
         assert catalog_endpoints.list("generated") is not None
-
-
-@pytest.mark.db_integration
-def test_projection_async(arrow_client: AuthenticatedArrowClient, query_runner: QueryRunner) -> None:
-    try:
-        endpoints = CatalogArrowEndpoints(arrow_client, query_runner)
-        handle = endpoints.project_async(
-            graph_name="g",
-            query="UNWIND range(1, 10) AS x WITH gds.graph.project.remote(x, null) as g RETURN g",
-        )
-
-        assert handle.job_id()
-
-        handle.wait()
-        assert handle.done()
-
-        G, result = handle.result()
-
-        assert G.name() == "g"
-        assert result["graphName"] == "g"
-        assert result["nodeCount"] == 10
-        assert result["relationshipCount"] == 0
-        assert result["projectMillis"] >= 0
-
-        assert len(endpoints.list("g")) == 1
-    finally:
-        endpoints.drop("g", fail_if_missing=False)
-
-
-@pytest.mark.db_integration
-def test_store_projection_async(arrow_client: AuthenticatedArrowClient, query_runner: QueryRunner) -> None:
-    try:
-        query_runner.run_cypher(
-            "UNWIND range(1, 5) AS x CREATE (:Person)-[:KNOWS]->(:Person)",
-            QueryType.USER_ACTION,
-        )
-
-        endpoints = CatalogArrowEndpoints(arrow_client, query_runner)
-        handle = endpoints.project_native_async(
-            graph_name="g",
-            node_label_filter=["Person"],
-            relationship_type_filter=["KNOWS"],
-        )
-
-        assert handle.job_id()
-
-        G, result = handle.result()
-
-        assert G.name() == "g"
-        assert result["graphName"] == "g"
-        assert result["nodeCount"] == 10
-        assert result["relationshipCount"] == 5
-
-        assert len(endpoints.list("g")) == 1
-    finally:
-        endpoints.drop("g", fail_if_missing=False)
-        query_runner.run_cypher("MATCH (n) DETACH DELETE n", QueryType.USER_ACTION)
 
 
 def test_graph_filter_async(catalog_endpoints: CatalogArrowEndpoints, sample_graph: Graph) -> None:
